# docs/cloud-api.openapi.yaml
openapi: 3.1.0
info:
  title: BaseBuddy Cloud API
  version: 1.0.0
  description: |
    Managed inference and training for BaseBuddy self-hosted installs.
    Proprietary service — OSS client in `core/inference/cloud/client.py`.
servers:
  - url: https://api.basebuddy.io/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    TrainingJobRequest:
      type: object
      required: [dataset_id, base_model]
      properties:
        dataset_id: { type: string, format: uuid }
        base_model: { type: string }
    TrainingJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        status:
          type: string
          enum: [queued, running, succeeded, failed]
        base_model: { type: string }
        result_model_id: { type: string, nullable: true }
        error: { type: string, nullable: true }
...
paths:
...
  /training/jobs:
    post:
      summary: Start a fine-tuning job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrainingJobRequest' }
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrainingJob' }
  /training/jobs/{job_id}:
    get:
      summary: Training job status
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TrainingJob' }
...